SBSA: Enhancements and bug fixes
Memory Map: EFI_NO_MAPPING macro removed from test_pool
- UEFI macro was present in test_pool/peripherals/test_m001.c
  This is a PAL representation and hence removed its reference
  from test_pool.

PCIe: Fix to add bdf 0x0 in table if type is RP

PCIe: BDF table is created before enumeration is performed in ACS
- This means that when creating the bdf table by interrogating the
  PCIe topology, the subordinate and secondary bus registers in
  the RC have not yet been programmed, and so accesses to all BDFs
  return 0xFFFF.

Exerciser: Identify Exerciser function based on BDF
- Exerciser is identified based on Device/Vendor ID
  and should be updated to identify based on BDF number.

PCIe: Moved pcie_is_onchip_peripheral() function from VAL to PAL
- This is an implementation defined function. Hence moving it
  from VAL to PAL.

Bug Fixes on SBSA
 - Added is_onchip_peripheral to platform Linux
 - Typo in defines used

Support for 64-bit ECAM address
 - RD-N2 has 64 bit ecam space. Generates exception in case of exerciser
   tests due to 32 bit ecam_base variable.
   Updated ecam_base to use a 64 bit variable.

Interrupt: Checking the status for val_pe_update_esr()
- Checking the status of this function help to identify if this
  was not implemented in baremetal platforms

Readme: Updated readme with a note for Baremetal.
- For baremetal PCIe enumeration, only the code provided with
  the ACS should be used.

Add status check for time delay function

PCIe: Adding link status polling in test 451
- Added check to see if the link status is down.

PCIe: Added RP type 1 BAR check in test 434

+++ val/src/avs_exerciser.c
@@ -150,7 +150,7 @@ uint32_t val_exerciser_get_state(EXERCISER_STATE *state, uint32_t instance)
 uint32_t val_exerciser_init(uint32_t instance)
 {
   uint32_t Bdf;
-  uint32_t Ecam;
+  uint64_t Ecam;
   uint64_t cfg_addr;
   EXERCISER_STATE state;
